Remove TranslatingVisitor.cast method that had no useful effect
This is an unchecked cast - it doesn't throw a ClassCastException if the
object has the wrong subclass of BValue. (It would throw an exception if
the object isn't a BValue at all, but in this code, it always is.) It's
just as safe (and simpler) to do a single unchecked cast at the end in
getResult.